requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
google-api-core==1.14.3
google-auth==1.7.0
google-cloud-core==1.0.3
google-cloud-firestore==1.5.0
google-cloud-storage==1.22.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.25.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
urllib3==1.25.6
Werkzeug==0.16.0
